+++ lib/samanage/api.rb
@@ -20,9 +20,8 @@ module Samanage
 
 		# Development mode forzes authorization & prepopulates custom forms/fields and admins
 		# datacenter should equal 'eu' or blank
-		def initialize(token: nil, datacenter: nil, development_mode: false, max_retries: MAX_RETRIES)
-			self.max_retries = max_retries
-			self.token = token if token
+		def initialize(token: , datacenter: nil, development_mode: false, max_retries: MAX_RETRIES)
+			self.token = token
 			if !datacenter.nil? && datacenter.to_s.downcase != 'eu'
 				datacenter = nil
 			end
@@ -30,6 +29,7 @@ module Samanage
 			self.base_url =  "https://api#{self.datacenter.to_s.downcase}.samanage.com/"
 			self.content_type = 'json'
 			self.admins = []
+			self.max_retries = max_retries
 			if development_mode
 				if self.authorized? != true
 					self.authorize