+++ lib/samanage/api/incidents.rb
@@ -14,8 +14,9 @@ module Samanage
 			incidents = Array.new
 			total_pages = self.get_incidents[:total_pages]
 			layout = options[:layout] == 'long' ? '&layout=long' : nil
+			archives = options[:audit_archives] ? '&audit_archives=true' : nil
 			while page <= total_pages
-				incidents += self.execute(http_method: 'get', path: "incidents.json?page=#{page}#{layout}")[:data]
+				incidents += self.execute(http_method: 'get', path: "incidents.json?page=#{page}#{layout}#{archives}")[:data]
 				page += 1
 			end
 			incidents