+++ lib/samanage/error.rb
@@ -1,8 +1,9 @@
 module Samanage
-	# Errors in gem
 	class Error < StandardError
 		attr_accessor :status_code, :response, :error
+		
 		def initialize(error: nil, response: {})
+			self.error = error
 			if response.class == Hash
 				self.status_code = response[:code]
 				self.response = response[:data] ||= response[:response]
@@ -11,10 +12,8 @@ module Samanage
 				self.response = response
 			end
 
-			self.error = error
 			puts "[Error] #{self.status_code}: #{self.response}"
 		end
-
 	end
 
 	# API Errors